+++ lib/appium_lib_core/driver.rb
@@ -276,6 +276,23 @@ module Appium
       #     @core = Appium::Core.for(opts) # create a core driver with 'opts' and extend methods into 'self'
       #     @core.start_driver # start driver with 'url'. Connect to 'http://custom-host:8080/wd/hub.com'
       #
+      #     # With a custom listener
+      #     class CustomListener < ::Selenium::WebDriver::Support::AbstractEventListener
+      #       // something
+      #     end
+      #     capabilities: {
+      #       platformName: :ios,
+      #       platformVersion: '11.0',
+      #       deviceName: 'iPhone Simulator',
+      #       automationName: 'XCUITest',
+      #       app: '/path/to/MyiOS.app'
+      #     },
+      #     appium_lib: {
+      #       listener: CustomListener.new,
+      #     }
+      #     @core = Appium::Core.for capabilities: capabilities, appium_lib: appium_lib
+      #     @core.start_driver
+      #
       def self.for(opts = {})
         new.setup_for_new_session(opts)
       end
@@ -488,13 +505,7 @@ module Appium
       end
 
       def get_http_client(http_client: nil, open_timeout: nil, read_timeout: nil)
-        client = http_client || Appium::Core::Base::Http::Default.new
-
-        # open_timeout and read_timeout are explicit wait.
-        client.open_timeout = open_timeout if open_timeout
-        client.read_timeout = read_timeout if read_timeout
-
-        client
+        http_client || Appium::Core::Base::Http::Default.new(open_timeout: open_timeout, read_timeout: read_timeout)
       end
 
       # Ignore setting default wait if the target driver has no implementation